Update README for for dev-hub parsing
In the dev-hub we are going to automatically sync the content from the cimg READMEs.
As a part of the copy in the dev-hub, the semantic versions in the Getting Started header are going to be updated to the latest tag.

https://circleci.com/developer/images/image/cimg/openjdk
Currently, the sample config is updated with the latest version semver, but the description below it mentions the text here.
As we automatically sync the content, we're simply going to update the semver in the entire header. This change makes the content in the header make more sense.
The explanation of the semver tags is in the Tagging Scheme header.